auth_project/token_auth_service/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
PyJWT==2.6.0